Refactoring
- removed unused rest dependency to plugin-core
- added plugin-core to framework
- moved mail functionality to framework
- moved application utility classes to framework
