+++ framework/src/main/java/fi/otavanopisto/pyramus/applications/DuplicatePersonException.java
@@ -1,0 +1,5 @@
+package fi.otavanopisto.pyramus.applications;
+
+public class DuplicatePersonException extends Exception {
+  private static final long serialVersionUID = 3871772751122159892L;
+}
