+++ framework/src/main/java/fi/otavanopisto/pyramus/mailer/MailEvent.java
@@ -1,4 +1,4 @@
-package fi.otavanopisto.pyramus.util;
+package fi.otavanopisto.pyramus.mailer;
 
 import java.util.List;
 import java.util.Set;
