+++ pyramus/pom.xml
@@ -392,13 +392,6 @@
       <artifactId>selenium-java</artifactId>
       <scope>test</scope>
     </dependency>
-
-    <dependency>
-      <groupId>javax.mail</groupId>
-      <artifactId>javax.mail-api</artifactId>
-      <version>1.6.0</version>
-      <scope>provided</scope>
-    </dependency>
     
     <!-- PDF (applications) -->
