+++ pyramus/src/main/java/fi/otavanopisto/pyramus/json/applications/ListMailTemplatesJSONRequestController.java
@@ -10,12 +10,12 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import fi.internetix.smvc.controllers.JSONRequestContext;
+import fi.otavanopisto.pyramus.applications.ApplicationUtils;
 import fi.otavanopisto.pyramus.dao.DAOFactory;
 import fi.otavanopisto.pyramus.dao.application.ApplicationMailTemplateDAO;
 import fi.otavanopisto.pyramus.domainmodel.application.ApplicationMailTemplate;
 import fi.otavanopisto.pyramus.framework.JSONRequestController;
 import fi.otavanopisto.pyramus.framework.UserRole;
-import fi.otavanopisto.pyramus.views.applications.ApplicationUtils;
 
 public class ListMailTemplatesJSONRequestController extends JSONRequestController {
