+++ pyramus/src/main/java/fi/otavanopisto/pyramus/json/applications/UpdateApplicationStateJSONRequestController.java
@@ -1,64 +1,31 @@
 package fi.otavanopisto.pyramus.json.applications;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import fi.internetix.smvc.controllers.JSONRequestContext;
+import fi.otavanopisto.pyramus.applications.ApplicationUtils;
+import fi.otavanopisto.pyramus.applications.DuplicatePersonException;
 import fi.otavanopisto.pyramus.dao.DAOFactory;
-import fi.otavanopisto.pyramus.dao.application.ApplicationAttachmentDAO;
 import fi.otavanopisto.pyramus.dao.application.ApplicationDAO;
 import fi.otavanopisto.pyramus.dao.application.ApplicationLogDAO;
 import fi.otavanopisto.pyramus.dao.application.ApplicationSignaturesDAO;
-import fi.otavanopisto.pyramus.dao.base.AddressDAO;
-import fi.otavanopisto.pyramus.dao.base.ContactTypeDAO;
-import fi.otavanopisto.pyramus.dao.base.EmailDAO;
 import fi.otavanopisto.pyramus.dao.base.PersonDAO;
-import fi.otavanopisto.pyramus.dao.base.PhoneNumberDAO;
-import fi.otavanopisto.pyramus.dao.file.StudentFileDAO;
-import fi.otavanopisto.pyramus.dao.students.StudentDAO;
-import fi.otavanopisto.pyramus.dao.system.SettingDAO;
-import fi.otavanopisto.pyramus.dao.system.SettingKeyDAO;
 import fi.otavanopisto.pyramus.dao.users.StaffMemberDAO;
-import fi.otavanopisto.pyramus.dao.users.UserDAO;
-import fi.otavanopisto.pyramus.dao.users.UserIdentificationDAO;
 import fi.otavanopisto.pyramus.domainmodel.application.Application;
-import fi.otavanopisto.pyramus.domainmodel.application.ApplicationAttachment;
 import fi.otavanopisto.pyramus.domainmodel.application.ApplicationLogType;
 import fi.otavanopisto.pyramus.domainmodel.application.ApplicationSignatureState;
 import fi.otavanopisto.pyramus.domainmodel.application.ApplicationSignatures;
 import fi.otavanopisto.pyramus.domainmodel.application.ApplicationState;
-import fi.otavanopisto.pyramus.domainmodel.base.ContactType;
-import fi.otavanopisto.pyramus.domainmodel.base.Email;
-import fi.otavanopisto.pyramus.domainmodel.base.Person;
-import fi.otavanopisto.pyramus.domainmodel.base.StudyProgramme;
-import fi.otavanopisto.pyramus.domainmodel.students.Sex;
 import fi.otavanopisto.pyramus.domainmodel.students.Student;
-import fi.otavanopisto.pyramus.domainmodel.system.Setting;
-import fi.otavanopisto.pyramus.domainmodel.system.SettingKey;
 import fi.otavanopisto.pyramus.domainmodel.users.StaffMember;
-import fi.otavanopisto.pyramus.domainmodel.users.User;
-import fi.otavanopisto.pyramus.domainmodel.users.UserIdentification;
 import fi.otavanopisto.pyramus.framework.JSONRequestController;
 import fi.otavanopisto.pyramus.framework.UserRole;
-import fi.otavanopisto.pyramus.plugin.auth.AuthenticationProviderVault;
-import fi.otavanopisto.pyramus.plugin.auth.InternalAuthenticationProvider;
-import fi.otavanopisto.pyramus.util.Mailer;
-import fi.otavanopisto.pyramus.views.applications.ApplicationUtils;
+import fi.otavanopisto.pyramus.mailer.Mailer;
 import net.sf.json.JSONObject;
 
 public class UpdateApplicationStateJSONRequestController extends JSONRequestController {
@@ -210,12 +177,12 @@ public class UpdateApplicationStateJSONRequestController extends JSONRequestCont
           
           // Separate logic for transferring the applicant as student
           
-          Student student = createPyramusStudent(application, staffMember);
+          Student student = ApplicationUtils.createPyramusStudent(application, staffMember); // throws exception if multiple persons found
           PersonDAO personDAO = DAOFactory.getInstance().getPersonDAO();
           personDAO.updateDefaultUser(student.getPerson(), student);
           String credentialToken = RandomStringUtils.randomAlphanumeric(32).toLowerCase();
           application = applicationDAO.updateApplicationStudentAndCredentialToken(application, student, credentialToken);
-          mailCredentialsInfo(requestContext, student, staffMember, application);
+          ApplicationUtils.mailCredentialsInfo(requestContext, student, staffMember, application);
         }
         
         // Update the actual application state
@@ -247,6 +214,11 @@ public class UpdateApplicationStateJSONRequestController extends JSONRequestCont
       requestContext.addResponseParameter("handlerId", application.getHandler() == null ? null : application.getHandler().getId());
       requestContext.addResponseParameter("lastModified", application.getLastModified().getTime());
     }
+    catch (DuplicatePersonException dpe) {
+      requestContext.addResponseParameter("status", "FAIL");
+      requestContext.addResponseParameter("reason", "Käyttäjätiedot täsmäävät useampaan olemassa olevaan käyttäjätiliin");
+      logger.log(Level.SEVERE, "Error updating application state", dpe);
+    }
     catch (Exception e) {
       requestContext.addResponseParameter("status", "FAIL");
       requestContext.addResponseParameter("reason", e.getMessage());
@@ -262,342 +234,9 @@ public class UpdateApplicationStateJSONRequestController extends JSONRequestCont
     requestContext.addResponseParameter("status", "FAIL");
     requestContext.addResponseParameter("reason", reason);
   }
-  
-  private Student createPyramusStudent(Application application, StaffMember staffMember) throws DuplicatePersonException {
-    PhoneNumberDAO phoneNumberDAO = DAOFactory.getInstance().getPhoneNumberDAO();
-    AddressDAO addressDAO = DAOFactory.getInstance().getAddressDAO();
-    EmailDAO emailDAO = DAOFactory.getInstance().getEmailDAO();
-    ContactTypeDAO contactTypeDAO = DAOFactory.getInstance().getContactTypeDAO();
-    PersonDAO personDAO = DAOFactory.getInstance().getPersonDAO();
-    StudentDAO studentDAO = DAOFactory.getInstance().getStudentDAO();
-    ApplicationAttachmentDAO applicationAttachmentDAO = DAOFactory.getInstance().getApplicationAttachmentDAO();
-    
-    JSONObject formData = JSONObject.fromObject(application.getFormData());
-    
-    // Create person (if needed)
-    
-    Person person = resolvePerson(application);
-    if (person == null) {
-      String birthdayStr = getFormValue(formData, "field-birthday");
-      String ssnEnd = getFormValue(formData, "field-ssn-end");
-      try {
-        Date birthday = StringUtils.isEmpty(birthdayStr) ? null : new SimpleDateFormat("d.M.yyyy").parse(birthdayStr);
-        String ssn = StringUtils.isBlank(ssnEnd) ? null : ApplicationUtils.constructSSN(birthdayStr, ssnEnd);
-        Sex sex = ApplicationUtils.resolveGender(getFormValue(formData, "field-sex"), ssnEnd);
-        person = personDAO.create(birthday, ssn, sex, null, Boolean.FALSE);
-      }
-      catch (ParseException e) {
-        logger.severe(String.format("Invalid birthday format in application entity %d", application.getId()));
-        return null;
-      }
-    }
-    
-    // Determine correct study programme
-    
-    StudyProgramme studyProgramme = ApplicationUtils.resolveStudyProgramme(
-        getFormValue(formData, "field-line"),
-        getFormValue(formData, "field-foreign-line"));
-    if (studyProgramme == null) {
-      logger.severe(String.format("Unable to resolve study programme of application entity %d", application.getId()));
-      return null;
-    }
-    
-    // Create student
-    
-    Student student = studentDAO.create(
-        person,
-        getFormValue(formData, "field-first-names"),
-        getFormValue(formData, "field-last-name"),
-        getFormValue(formData, "field-nickname"),
-        null, // additionalInfo,
-        null, // studyTimeEnd (TODO can this be resolved?)
-        ApplicationUtils.resolveStudentActivityType(getFormValue(formData, "field-job")),
-        ApplicationUtils.resolveStudentExaminationType(getFormValue(formData, "field-internetix-contract-school-degree")),
-        null, // student educational level (entity)
-        null, // education (string)
-        ApplicationUtils.resolveNationality(getFormValue(formData, "field-nationality")),
-        ApplicationUtils.resolveMunicipality(getFormValue(formData, "field-municipality")),
-        ApplicationUtils.resolveLanguage(getFormValue(formData, "field-language")),
-        ApplicationUtils.resolveSchool(getFormValue(formData, "field-internetix-contract-school")),
-        studyProgramme,
-        null, // curriculum (TODO can this be resolved?)
-        null, // previous studies (double)
-        new Date(), // study start date
-        null, // study end date
-        null, // study end reason
-        null, // study end text
-        Boolean.FALSE); // archived
-    
-    // Main contact type
-    
-    ContactType contactType = contactTypeDAO.findById(1L); // Koti (unique)
-
-    // Attach email
-    
-    String email = getFormValue(formData, "field-email");
-    emailDAO.create(student.getContactInfo(), contactType, Boolean.TRUE, email.toLowerCase());
-    
-    // Attach address
-    
-    addressDAO.create(
-        student.getContactInfo(),
-        contactType,
-        String.format("%s %s", getFormValue(formData, "field-nickname"), getFormValue(formData, "field-last-name")),
-        getFormValue(formData, "field-street-address"),
-        getFormValue(formData, "field-zip-code"),
-        getFormValue(formData, "field-city"),
-        getFormValue(formData, "field-country"),
-        Boolean.TRUE);
-
-    // Attach phone
-    
-    phoneNumberDAO.create(
-        student.getContactInfo(),
-        contactType,
-        Boolean.TRUE,
-        getFormValue(formData, "field-phone"));
-    
-    // Guardian info (if email is present, all other fields are required and present, too)
-    
-    email = getFormValue(formData, "field-underage-email");
-    if (!StringUtils.isBlank(email)) {
-      
-      // Attach email
-      
-      contactType = contactTypeDAO.findById(5L); // Yhteyshenkilö (non-unique)
-      emailDAO.create(student.getContactInfo(), contactType, Boolean.FALSE, email.toLowerCase());
-
-      // Attach address
-      
-      addressDAO.create(
-          student.getContactInfo(),
-          contactType,
-          String.format("%s %s", getFormValue(formData, "field-underage-first-name"), getFormValue(formData, "field-underage-last-name")),
-          getFormValue(formData, "field-underage-street-address"),
-          getFormValue(formData, "field-underage-zip-code"),
-          getFormValue(formData, "field-underage-city"),
-          getFormValue(formData, "field-underage-country"),
-          Boolean.FALSE);
-
-      // Attach phone
-      
-      phoneNumberDAO.create(
-          student.getContactInfo(),
-          contactType,
-          Boolean.FALSE,
-          getFormValue(formData, "field-underage-phone"));
-    }
-    
-    // Attachments
-    
-    List<ApplicationAttachment> attachments = applicationAttachmentDAO.listByApplicationId(application.getApplicationId());
-    if (!attachments.isEmpty()) {
-      String attachmentsFolder = getSettingValue("applications.storagePath");
-      if (StringUtils.isNotEmpty(attachmentsFolder)) {
-        StudentFileDAO studentFileDAO = DAOFactory.getInstance().getStudentFileDAO();
-        String applicationId = sanitizeFilename(application.getApplicationId());
-        for (ApplicationAttachment attachment : attachments) {
-          String attachmentFileName = sanitizeFilename(attachment.getName());
-          try {
-            java.nio.file.Path path = Paths.get(attachmentsFolder, applicationId, attachmentFileName);
-            File file = path.toFile();
-            if (file.exists()) {
-              String contentType = Files.probeContentType(path);
-              byte[] data = FileUtils.readFileToByteArray(file);
-              studentFileDAO.create(student, attachmentFileName, attachmentFileName, null, contentType, data, staffMember);
-            }
-          }
-          catch (IOException e) {
-            logger.log(Level.WARNING, String.format("Exception processing attachment %s of application %d",
-                attachment.getName(), application.getId()), e);
-          }
-        }
-      }
-    }
-    
-    return student;
-  }
-  
-  private Person resolvePerson(Application application) throws DuplicatePersonException {
-    EmailDAO emailDAO = DAOFactory.getInstance().getEmailDAO();
-    PersonDAO personDAO = DAOFactory.getInstance().getPersonDAO();
-    UserDAO userDAO = DAOFactory.getInstance().getUserDAO();
-    JSONObject applicationData = JSONObject.fromObject(application.getFormData());
-    
-    // Person by social security number
-    
-    Map<Long, Person> existingPersons = new HashMap<Long, Person>();
-    boolean hasSsn = getFormValue(applicationData, "field-ssn-end") != null;
-    if (hasSsn) {
-      
-      // Given SSN
-      
-      String ssn = ApplicationUtils.constructSSN(getFormValue(applicationData, "field-birthday"), getFormValue(applicationData, "field-ssn-end"));
-      List<Person> persons = personDAO.listBySSNUppercase(ssn);
-      for (Person person : persons) {
-        existingPersons.put(person.getId(), person);
-      }
-      
-      // SSN with "wrong" delimiter
-      
-      char[] ssnChars = ssn.toCharArray();
-      ssnChars[6] = ssnChars[6] == 'A' ? '-' : 'A';
-      ssn = ssnChars.toString();
-      persons = personDAO.listBySSNUppercase(ssn);
-      for (Person person : persons) {
-        existingPersons.put(person.getId(), person);
-      }
-    }
-    
-    // Person by email address
-    
-    String emailAddress = StringUtils.lowerCase(StringUtils.trim(application.getEmail()));
-    List<Email> emails = emailDAO.listByAddressLowercase(emailAddress);
-    for (Email email : emails) {
-      if (email.getContactType() != null && Boolean.FALSE.equals(email.getContactType().getNonUnique())) {
-        User user = userDAO.findByContactInfo(email.getContactInfo());
-        if (user != null) {
-          Person person = user.getPerson();
-          if (person != null) {
-            existingPersons.put(person.getId(), person);
-          }
-        }
-      }
-    }
-
-    if (existingPersons.size() > 1) {
-      StringBuilder sb = new StringBuilder();
-      for (Person person : existingPersons.values()) {
-        if (sb.length() > 0) {
-          sb.append(",");
-        }
-        sb.append(person.getId());
-      }
-      logger.severe(String.format("Persons with duplicate SSN or unique email: %s", sb));
-      throw new DuplicatePersonException();
-    }
-    else if (existingPersons.isEmpty()) {
-      return null;
-    }
-    else {
-      return existingPersons.values().iterator().next();
-    }
-  }
-  
-  private void mailCredentialsInfo(JSONRequestContext requestContext, Student student, StaffMember staffMember,
-      Application application) throws Exception {
-    
-    // Application form 
-    
-    JSONObject formData = JSONObject.fromObject(application.getFormData());
-    
-    // Determine the need for credentials
-    
-    List<InternalAuthenticationProvider> providers = AuthenticationProviderVault.getInstance().getInternalAuthenticationProviders();
-    InternalAuthenticationProvider provider = providers.size() == 1 ? providers.get(0) : null;
-    if (provider == null) {
-      throw new InternalError("Unable to resolve InternalAuthenticationProvider");
-    }
-    UserIdentificationDAO userIdentificationDAO = DAOFactory.getInstance().getUserIdentificationDAO();
-    UserIdentification userIdentification = userIdentificationDAO.findByAuthSourceAndPerson(provider.getName(), student.getPerson());
-    
-    // Choose mail template depending on whether student already has credentials or not
-    
-    String subject = "Muikku-oppimisympäristön tunnukset";
-    String content;
-    if (userIdentification == null) {
-      StringBuilder createCredentialsUrl = new StringBuilder();
-      createCredentialsUrl.append(requestContext.getRequest().getScheme());
-      createCredentialsUrl.append("://");
-      createCredentialsUrl.append(requestContext.getRequest().getServerName());
-      createCredentialsUrl.append(":");
-      createCredentialsUrl.append(requestContext.getRequest().getServerPort());
-      createCredentialsUrl.append("/applications/createcredentials.page?a=");
-      createCredentialsUrl.append(application.getApplicationId());
-      createCredentialsUrl.append("&t=");
-      createCredentialsUrl.append(application.getCredentialToken());
-      
-      content = IOUtils.toString(requestContext.getServletContext().getResourceAsStream(
-          "/templates/applications/mail-credentials-create.html"), "UTF-8");
-      content = String.format(
-          content,
-          getFormValue(formData, "field-nickname"),
-          createCredentialsUrl,
-          staffMember.getFullName());
-    }
-    else {
-      content = IOUtils.toString(requestContext.getServletContext().getResourceAsStream(
-          "/templates/applications/mail-credentials-exist.html"), "UTF-8");
-      content = String.format(
-          content,
-          getFormValue(formData, "field-nickname"),
-          staffMember.getFullName());
-    }
-      
-    // Send mail to applicant (and possible guardian)
-    
-    if (StringUtils.isBlank(getFormValue(formData, "field-underage-email"))) {
-      Mailer.sendMail(
-          Mailer.JNDI_APPLICATION,
-          Mailer.HTML,
-          null,
-          application.getEmail(),
-          subject,
-          content);
-    }
-    else {
-      Mailer.sendMail(
-          Mailer.JNDI_APPLICATION,
-          Mailer.HTML,
-          null,
-          application.getEmail(),
-          getFormValue(formData, "field-underage-email"),
-          subject,
-          content);
-    }
-    
-    // Add notification about sent mail
-    
-    ApplicationLogDAO applicationLogDAO = DAOFactory.getInstance().getApplicationLogDAO();
-    applicationLogDAO.create(
-      application,
-      ApplicationLogType.HTML,
-      String.format("<p>%s</p><p><b>%s</b></p>%s", "Hakijalle lähetetty ohjeet Muikku-tunnuksista", subject, content),
-      staffMember);
-  }
 
   private String getFormValue(JSONObject object, String key) {
     return object.has(key) ? object.getString(key) : null;
   }
-
-  private String getSettingValue(String key) {
-    SettingKeyDAO settingKeyDAO = DAOFactory.getInstance().getSettingKeyDAO();
-    SettingKey settingKey = settingKeyDAO.findByName(key);
-    if (settingKey != null) {
-      SettingDAO settingDAO = DAOFactory.getInstance().getSettingDAO();
-      Setting setting = settingDAO.findByKey(settingKey);
-      if (setting != null && setting.getValue() != null) {
-        return setting.getValue();
-      }
-    }
-    return null;
-  }
-
-  /**
-   * Sanitizes the given filename so that it can be safely used as part of a file path. The filename
-   * is first stripped of traditional invalid filename characters (\ / : * ? " < > |) and then of all
-   * leading and trailing periods.
-   * 
-   * @param filename Filename to be sanitized
-   * 
-   * @return Sanitized filename
-   */
-  private String sanitizeFilename(String filename) {
-    return StringUtils.stripEnd(StringUtils.stripStart(StringUtils.strip(filename, "\\/:*?\"<>|"), "."), ".");
-  }
   
-  private class DuplicatePersonException extends Exception {
-    private static final long serialVersionUID = -1840064717234172676L;
-  }
-
 }
