+++ rest/pom.xml
@@ -105,11 +105,6 @@
     </dependency>
 
     <dependency>
-      <groupId>fi.otavanopisto.pyramus</groupId>
-      <artifactId>plugin-core</artifactId>
-    </dependency>
-
-    <dependency>
       <groupId>com.fasterxml.jackson.datatype</groupId>
       <artifactId>jackson-datatype-jsr310</artifactId>
     </dependency>
